[Fix] #130 - 폴더링 변경 및 불필요한 파일 삭제
