+++ Sopetit-iOS/Sopetit-iOS/Presentation/Achieve/AchieveScene/View/AchieveView.swift
@@ -1,0 +1,42 @@
+//
+//  AchieveView.swift
+//  Sopetit-iOS
+//
+//  Created by 고아라 on 5/27/24.
+//
+
+import UIKit
+
+import SnapKit
+
+final class AchieveView: UIView {
+    
+    // MARK: - UI Components
+    
+    // MARK: - Life Cycles
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        setHierarchy()
+        setLayout()
+    }
+    
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
+// MARK: - Extensions
+
+private extension AchieveView {
+    
+    func setHierarchy() {
+        
+    }
+    
+    func setLayout() {
+        
+    }
+}
