+++ Sopetit-iOS/Sopetit-iOS/Presentation/Achieve/AchieveScene/ViewController/AchieveViewController.swift
@@ -1,0 +1,38 @@
+//
+//  AchieveViewController.swift
+//  Sopetit-iOS
+//
+//  Created by 고아라 on 5/27/24.
+//
+
+import UIKit
+
+import SnapKit
+
+final class AchieveViewController: UIViewController {
+    
+    // MARK: - UI Components
+    
+    private var achieveView = AchieveView()
+    
+    // MARK: - Life Cycles
+    
+    override func loadView() {
+        self.view = achieveView
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        setUI()
+    }
+}
+
+// MARK: - Extensions
+
+extension AchieveViewController {
+    
+    func setUI() {
+        self.navigationController?.navigationBar.isHidden = true
+    }
+}
